pool: move `RelayPoolOptions` to `pool::options` module
