+++ crates/nostr-relay-pool/src/pool/options.rs
